requirements.txt
loguru
MetaTrader5
pandas
psutil
pydantic
pydantic-settings
python-telegram-bot
pytest
